# backend/docs/swagger.yaml
basePath: /
definitions:
...
  do.ChannelDTO:
    properties:
      credentials:
        additionalProperties:
          type: string
        type: object
      enabled:
        type: boolean
      extra:
        additionalProperties:
          type: string
        type: object
      id:
        example: "0"
        type: string
      name:
        type: string
      owner_id:
        example: "0"
        type: string
      owner_scope:
        type: string
      type:
        type: string
    type: object
  do.CreateChannelRequest:
    properties:
      credentials:
        additionalProperties:
          type: string
        type: object
      enabled:
        type: boolean
      extra:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      type:
        type: string
    type: object
...
  do.CreatePushChannelRequest:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      other:
        type: string
      token:
        type: string
      type:
        type: string
      url:
        type: string
    required:
    - name
    - type
    type: object
...
  do.Definition:
    properties:
      fields:
        items:
          $ref: '#/definitions/do.Field'
        type: array
      type:
        type: string
    type: object
  do.Field:
    properties:
      key:
        type: string
      required:
        type: boolean
      type:
        type: string
    type: object
...
  do.TestPushChannelRequest:
    properties:
      name:
        type: string
      other:
        type: string
      target:
        type: string
      token:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
...
  do.UpdateChannelRequest:
    properties:
      credentials:
        additionalProperties:
          type: string
        type: object
      enabled:
        type: boolean
      extra:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    type: object
  do.UpdatePushChannelRequest:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      other:
        type: string
      token:
        type: string
      type:
        type: string
      url:
        type: string
    required:
    - type
    type: object
...
  entity.PushChannel:
    properties:
      created_at:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        example: "0"
        type: string
      name:
        type: string
      other:
        type: string
      token:
        type: string
      type:
        type: string
      updated_at:
        type: string
      url:
        type: string
    type: object
...
  response.Any:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
...
info:
  contact:
    name: Wavelet
    url: https://github.com/Rain-kl/Wavelet
  description: Wavelet 平台后端 API，提供用户认证、系统配置、任务调度等通用功能。
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Wavelet API
  version: 1.0.0
paths:
...
  /api/v1/admin/message-gateway/channels:
    get:
      description: Returns all messaging channels; secrets are masked
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/do.ChannelDTO'
                  type: array
              type: object
      security:
      - SessionCookie: []
      summary: List message gateway channels
      tags:
      - admin-message-gateway
    post:
      consumes:
      - application/json
      description: Creates a Telegram or QQ channel with encrypted credentials
      parameters:
      - description: create body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.CreateChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/do.ChannelDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: Create message gateway channel
      tags:
      - admin-message-gateway
  /api/v1/admin/message-gateway/channels/{id}:
    delete:
      description: Deletes a channel and cascaded bindings and pairing codes
      parameters:
      - description: channel id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: Delete message gateway channel
      tags:
      - admin-message-gateway
    patch:
      consumes:
      - application/json
      description: Updates a channel; empty secrets keep the current ciphertext
      parameters:
      - description: channel id
        in: path
        name: id
        required: true
        type: integer
      - description: update body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.UpdateChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/do.ChannelDTO'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: Update message gateway channel
      tags:
      - admin-message-gateway
  /api/v1/admin/message-gateway/channels/{id}/test:
    post:
      description: Probes stored credentials without returning secrets
      parameters:
      - description: channel id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Any'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: Test message gateway channel
      tags:
      - admin-message-gateway
  /api/v1/admin/message-gateway/channels/definitions:
    get:
      description: Returns form field definitions for Telegram and QQ channels
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/do.Definition'
                  type: array
              type: object
      security:
      - SessionCookie: []
      summary: List message gateway channel definitions
      tags:
      - admin-message-gateway
  /api/v1/admin/push/channels:
    get:
      description: 返回系统配置的所有消息通道列表，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 消息通道列表
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.PushChannel'
                  type: array
              type: object
      security:
      - SessionCookie: []
      summary: 获取所有消息通道
      tags:
      - admin-push
    post:
      consumes:
      - application/json
      description: 新建一个消息通道配置，需要管理员权限
      parameters:
      - description: 创建参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.CreatePushChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/entity.PushChannel'
              type: object
      security:
      - SessionCookie: []
      summary: 创建消息通道
      tags:
      - admin-push
  /api/v1/admin/push/channels/{id}:
    delete:
      description: 根据ID删除消息通道，需要管理员权限
      parameters:
      - description: 通道ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 删除消息通道
      tags:
      - admin-push
    put:
      consumes:
      - application/json
      description: 修改消息通道配置，需要管理员权限
      parameters:
      - description: 通道ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.UpdatePushChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Any'
            - properties:
                data:
                  $ref: '#/definitions/entity.PushChannel'
              type: object
      security:
      - SessionCookie: []
      summary: 更新消息通道
      tags:
      - admin-push
  /api/v1/admin/push/channels/definitions:
    get:
      description: 返回系统支持的所有消息通道类型的动态表单定义，需要管理员权限
      produces:
      - application/json
      responses:
        "200":
          description: 通道配置定义列表
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 获取所有消息通道配置字段定义
      tags:
      - admin-push
  /api/v1/admin/push/channels/test:
    post:
      consumes:
      - application/json
      description: 触发一次临时的或现有的通道连通性推送测试，需要管理员权限
      parameters:
      - description: 测试参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/do.TestPushChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 测试触发成功
          schema:
            $ref: '#/definitions/response.Any'
      security:
      - SessionCookie: []
      summary: 测试通道连通性
      tags:
      - admin-push
...
securityDefinitions:
  SessionCookie:
    in: cookie
    name: session
    type: apiKey
swagger: "2.0"